+++ docs/source/augmentation.container.rst
@@ -3,7 +3,7 @@ Augmentation Containers
 
 .. currentmodule:: kornia.augmentation.container
 
-The classes in this section are containers for augmenting different data formats (e.g. videos).
+The classes in this section are containers for augmenting different data formats (e.g. images, videos).
 
 
 Augmentation Sequential
@@ -11,11 +11,25 @@ Augmentation Sequential
 
 Kornia augmentations provides simple on-device augmentation framework with the support of various syntax sugars
 (e.g. return transformation matrix, inverse geometric transform). Therefore, we provide advanced augmentation
-container to ease the pain of building augmenation pipelines.
+container to ease the pain of building augmenation pipelines. This API would also provide predefined routines
+for automating the processing of masks, bounding boxes, and keypoints.
 
 .. autoclass:: AugmentationSequential
 
+   .. automethod:: forward
+
    .. automethod:: inverse
+
+
+
+ImageSequential
+-----------------------
+   
+Kornia augmentations provides simple on-device augmentation framework with the support of various syntax sugars
+(e.g. return transformation matrix, inverse geometric transform). Additionally, ImageSequential supports the
+mix usage of both image processing and augmentation modules.
+
+.. autoclass:: ImageSequential
 
    .. automethod:: forward
